const in one place
