+++ calyx-py/calyx/fifo_oracle.py
@@ -1,8 +1,8 @@
 import queues
-
+import queue_util
 
 if __name__ == "__main__":
-    commands, values = queues.parse_json()
+    commands, values = queue_util.parse_json()
     pifo = queues.Fifo([])
     ans = queues.operate_queue(commands, values, pifo)
-    queues.dump_json(commands, values, ans)
+    queue_util.dump_json(commands, values, ans)
