+++ calyx-py/calyx/queue_call.py
@@ -1,7 +1,6 @@
 # pylint: disable=import-error
+import calyx.queue_util as queue_util
 import calyx.builder as cb
-
-MAX_CMDS = 15
 
 
 def insert_main(prog, queue):
@@ -32,9 +31,9 @@ def insert_main(prog, queue):
     # - one ref register, `ans`, into which the result of a pop or peek is written.
     # - one ref register, `err`, which is raised if an error occurs.
 
-    commands = main.seq_mem_d1("commands", 2, MAX_CMDS, 32, is_external=True)
-    values = main.seq_mem_d1("values", 32, MAX_CMDS, 32, is_external=True)
-    ans_mem = main.seq_mem_d1("ans_mem", 32, MAX_CMDS, 32, is_external=True)
+    commands = main.seq_mem_d1("commands", 2, queue_util.MAX_CMDS, 32, is_external=True)
+    values = main.seq_mem_d1("values", 32, queue_util.MAX_CMDS, 32, is_external=True)
+    ans_mem = main.seq_mem_d1("ans_mem", 32, queue_util.MAX_CMDS, 32, is_external=True)
 
     # The two components we'll use:
     queue = main.cell("myqueue", queue)
